Revert "net: sockets: move poll implementation to zvfs"
This reverts commit 93973e2ead9bc1e8e1bcf05d46cc292a05b49fdd.

PR #73978 introduced a regression.
Unfortunately this PR cannot be reverted without reverting also
Let's revert both PRs to stabilize main again towards the 3.7 release.

For more details on the issue see
https://github.com/zephyrproject-rtos/zephyr/issues/75205
