+++ lib/os/zvfs/CMakeLists.txt
@@ -2,4 +2,3 @@
 
 zephyr_library()
 zephyr_library_sources_ifdef(CONFIG_ZVFS_EVENTFD zvfs_eventfd.c)
-zephyr_library_sources_ifdef(CONFIG_ZVFS_POLL zvfs_poll.c)
