+++ lib/posix/options/Kconfig.device_io
@@ -6,10 +6,9 @@ menu "POSIX device I/O"
 
 config POSIX_DEVICE_IO
 	bool "POSIX device I/O [EXPERIMENTAL]"
+	select FDTABLE
 	select EXPERIMENTAL
 	select REQUIRES_FULL_LIBC
-	select ZVFS
-	select ZVFS_POLL
 	help
 	  Select 'y' here and Zephyr will provide an implementation of the POSIX_DEVICE_IO Option
 	  Group such as FD_CLR(), FD_ISSET(), FD_SET(), FD_ZERO(), close(), fdopen(), fileno(), open(),
