+++ subsys/net/lib/sockets/Kconfig
@@ -5,8 +5,7 @@
 
 menuconfig NET_SOCKETS
 	bool "BSD Sockets compatible API"
-	select ZVFS
-	select ZVFS_POLL
+	select FDTABLE
 	help
 	  Provide BSD Sockets like API on top of native Zephyr networking API.
 
@@ -48,7 +47,9 @@ config NET_SOCKETS_POSIX_NAMES
 
 config NET_SOCKETS_POLL_MAX
 	int "Max number of supported poll() entries"
-	default ZVFS_POLL_MAX
+	default 6 if WIFI_NM_WPA_SUPPLICANT
+	default 4 if SHELL_BACKEND_TELNET
+	default 3
 	help
 	  Maximum number of entries supported for poll() call.
