+++ tests/net/lib/lwm2m/lwm2m_engine/prj.conf
@@ -1,5 +1,3 @@
 CONFIG_ZTEST=y
 CONFIG_ZTEST_STACK_SIZE=5120
 CONFIG_MP_MAX_NUM_CPUS=1
-
-CONFIG_ZVFS=y
